Added the Base for displaying section headers. Need to refactor in order for it to work. But atleast it's not crashing and it's working

+++ app/src/main/java/com/pedrofr/sportsfinder/data/model/Odd.kt
@@ -6,7 +6,7 @@ import java.util.*
 
 @Entity(tableName = "Odd")
 data class Odd(
-    @PrimaryKey(autoGenerate = true) val oddsKey: Int = 0,
+    @PrimaryKey(autoGenerate = true) val oddsKey: Long = 0,
     val sportsKey: String,
     val startTime: String,
     val homeTeam: String,
